Fixed a wrong type being used.

+++ src/faction.c
@@ -697,7 +697,7 @@ int faction_isPlayerFriend( int f )
    Faction *faction;
    lua_State *L;
    int errf;
-   double r;
+   int r;
 
    faction = &faction_stack[f];
 
@@ -761,7 +761,7 @@ int faction_isPlayerEnemy( int f )
    Faction *faction;
    lua_State *L;
    int errf;
-   double r;
+   int r;
 
    faction = &faction_stack[f];
